Get Patches from 3D Volume

Get Patches from 3D Volume

get_patch_from_volume(vol, mask = NULL, patchsize, verbose = TRUE,
  pad = TRUE, normalize = TRUE, contrast, seed = NULL,
  run_mask_patches = TRUE, peak_estimator = c("Python", "R"))

get_num_patches(mask)

norm_pad(vol, mask = NULL, patchsize, verbose = TRUE, pad = TRUE,
  normalize = TRUE, peak_estimator = c("Python", "R"), contrast)

get_mask_patches(mask, patchsize, pad = TRUE, seed = NULL,
  verbose = TRUE)

Arguments

vol
3D array or nifti image

mask
binary 3D array or nifti image, for the brain usually

patchsize
Vector of length 2 (or more)

verbose
print diagnostic messages

pad
Run pad_image on the image before getting patches (pads then normalizes if normalize = TRUE)

normalize
Run normalize_image on the image before getting patches

contrast
What imaging sequence of MRI is this volume, passed to normalize_image

seed
Seed for random sampling of indices. If NULL, no sampling is done, passed to mask_indices

run_mask_patches
Should the patches for the mask be made as well?

peak_estimator
Which functions to use to estimate peak, either wrapper Python code or native R code.

...
not used

Value

A list of image and mask Patches

Examples

user = Sys.getenv("USER")
if (user == "johnmuschelli") {
reticulate::use_python(paste0(
"/Library/Frameworks/Python.framework/Versions/3.5/bin/python3"))
mpl = reticulate::import("matplotlib")
mpl$use('TkAgg')
}
library(neurobase)

#> Loading required package: oro.nifti

#> oro.nifti 0.10.1

fname = system.file("extdata", "MPRAGE.nii.gz", package = "flexconn")
vol = readnii(fname)
mask = vol > 0
patchsize = c(5, 5)
verbose = TRUE
patch = get_patch_from_volume(vol, mask, patchsize = patchsize,
contrast = "T1")

#> 99th quantile is 1136581.056, bandwidth =14207.263, gridsize =80

#> 4 peaks found.

#> Peak found at 1044512.47383877 for T1

#> Size matrix to create: 1596366x5x5x1

#> Size matrix to create: 1596366x5x5x1

rm(patch)

patch = get_patch_from_volume(vol, mask, patchsize = c(1,1,1),
contrast = "T1")

#> 99th quantile is 1136581.056, bandwidth =14207.263, gridsize =80

#> 4 peaks found.

#> Peak found at 1044512.47383877 for T1

#> Size matrix to create: 1596366x1x1x1x1

#> Size matrix to create: 1596366x1x1x1x1

rm(patch)